+++ src/lib/helpers/hashed.js
