+++ src/lib/index.js
@@ -1,5 +1,5 @@
 import AuthContext  from './components/auth-context'
 import {Authenticated}  from './components/authenticated'
-import { getLocalToken } from '../local-token'
+import { getLocalToken } from './local-token'
 
 export {AuthContext,Authenticated,getLocalToken}
